WSADataPortal/requirements.txt
BeautifulSoup==3.2.1
beautifulsoup4==4.4.1
bs4==0.0.1
demjson==2.2.4
dj-database-url==0.4.1
Django==1.9.6
djangorestframework==3.3.3
gunicorn==19.5.0
Pygments==2.1.3
whitenoise==3.1
